+++ sarasvant/lib/router/router.dart
@@ -7,7 +7,7 @@ import '../apps/teacher/screens/add_question/add_question.dart';
 import '../apps/teacher/screens/classrooms/classrooms.dart';
 import '../apps/teacher/screens/dashboard/dashboard.dart';
 import '../apps/teacher/screens/question_bank/question_bank.dart';
-import '../apps/teacher/screens/questions/questions.dart';
+import '../apps/teacher/screens/questions/desktop/questions.dart';
 
 final GoRouter router = GoRouter(
   initialLocation: '/teacher-dashboard',
